+++ zone/bot.cpp
@@ -509,7 +509,7 @@ void Bot::SetBotID(uint32 botID) {
 	npctype_id = botID;
 }
 
-void Bot::SetBotSpellID(uint32 newSpellID) {
+void Bot::SetBotSpellID(int32 newSpellID) {
 	npc_spells_id = newSpellID;
 }
 
@@ -1542,7 +1542,7 @@ bool Bot::LoadPet()
 	if (!pet_index)
 		return true;
 
-	uint32 saved_pet_spell_id = 0;
+	int32 saved_pet_spell_id = 0;
 	database.botdb.LoadPetSpellID(GetBotID(), saved_pet_spell_id);
 
 	if (!IsValidSpell(saved_pet_spell_id)) {
@@ -1554,7 +1554,7 @@ bool Bot::LoadPet()
 	std::string pet_name;
 	uint32 pet_mana = 0;
 	uint32 pet_hp = 0;
-	uint32 pet_spell_id = 0;
+	int32 pet_spell_id = 0;
 
 	if (!database.botdb.LoadPetStats(GetBotID(), pet_name, pet_mana, pet_hp, pet_spell_id)) {
 		return false;
@@ -2321,7 +2321,7 @@ void Bot::AI_Process()
 			}
 
 			if (!at_combat_range && RuleB(Bots, UseSpellPulling)) {
-				uint16 pull_spell_id = RuleI(Bots, PullSpellID);
+				int32 pull_spell_id = RuleI(Bots, PullSpellID);
 
 				if (IsValidSpell(pull_spell_id) && tar_distance <= spells[pull_spell_id].range) {
 					at_combat_range = true;
@@ -2363,7 +2363,7 @@ void Bot::AI_Process()
 				}
 
 				if (RuleB(Bots, UseSpellPulling)) {
-					uint16 pull_spell_id = RuleI(Bots, PullSpellID);
+					int32 pull_spell_id = RuleI(Bots, PullSpellID);
 
 					if (IsValidSpell(pull_spell_id) && tar_distance <= spells[pull_spell_id].range) {
 						StopMoving(CalculateHeadingToTarget(tar->GetX(), tar->GetY()));
@@ -4920,7 +4920,7 @@ void Bot::PerformTradeWithClient(int16 begin_slot_id, int16 end_slot_id, Client*
 	}
 }
 
-bool Bot::Death(Mob *killer_mob, int64 damage, uint16 spell_id, EQ::skills::SkillType attack_skill, KilledByTypes killed_by, bool is_buff_tic)
+bool Bot::Death(Mob *killer_mob, int64 damage, int32 spell_id, EQ::skills::SkillType attack_skill, KilledByTypes killed_by, bool is_buff_tic)
 {
 	if (!NPC::Death(killer_mob, damage, spell_id, attack_skill)) {
 		return false;
@@ -4966,7 +4966,7 @@ bool Bot::Death(Mob *killer_mob, int64 damage, uint16 spell_id, EQ::skills::Skil
 	return true;
 }
 
-void Bot::Damage(Mob *from, int64 damage, uint16 spell_id, EQ::skills::SkillType attack_skill, bool avoidable, int8 buffslot, bool iBuffTic, eSpecialAttacks special) {
+void Bot::Damage(Mob *from, int64 damage, int32 spell_id, EQ::skills::SkillType attack_skill, bool avoidable, int8 buffslot, bool iBuffTic, eSpecialAttacks special) {
 	if (!from) {
 		return;
 	}
@@ -5598,7 +5598,7 @@ void Bot::DoClassAttacks(Mob *target, bool IsRiposte) {
 	classattack_timer.Start(reuse / HasteModifier);
 }
 
-void Bot::MakePet(uint16 spell_id, const char* pettype, const char *petname) {
+void Bot::MakePet(int32 spell_id, const char* pettype, const char *petname) {
 	Mob::MakePet(spell_id, pettype, petname);
 }
 
@@ -5800,7 +5800,7 @@ void Bot::SetAttackTimer() {
 	}
 }
 
-int32 Bot::GetActSpellDuration(uint16 spell_id, int32 duration) {
+int32 Bot::GetActSpellDuration(int32 spell_id, int32 duration) {
 	int increase = 100;
 	increase += GetFocusEffect(focusSpellDuration, spell_id);
 	int64 tic_inc = 0;
@@ -5832,7 +5832,7 @@ int32 Bot::GetActSpellDuration(uint16 spell_id, int32 duration) {
 	return (((duration * increase) / 100) + tic_inc);
 }
 
-float Bot::GetAOERange(uint16 spell_id) {
+float Bot::GetAOERange(int32 spell_id) {
 	float range;
 	range = spells[spell_id].aoe_range;
 	if (range == 0)
@@ -5849,7 +5849,7 @@ float Bot::GetAOERange(uint16 spell_id) {
 	return range;
 }
 
-bool Bot::SpellEffect(Mob* caster, uint16 spell_id, float partial) {
+bool Bot::SpellEffect(Mob* caster, int32 spell_id, float partial) {
 	bool Result = false;
 	Result = Mob::SpellEffect(caster, spell_id, partial);
 	if (IsGrouped()) {
@@ -5871,7 +5871,7 @@ void Bot::DoBuffTic(const Buffs_Struct &buff, int slot, Mob* caster) {
 }
 
 bool Bot::CastSpell(
-	uint16 spell_id,
+	int32 spell_id,
 	uint16 target_id,
 	EQ::spells::CastingSlot slot,
 	int32 cast_time,
@@ -5951,7 +5951,7 @@ bool Bot::CastSpell(
 }
 
 bool Bot::SpellOnTarget(
-		uint16 spell_id,
+		int32 spell_id,
 		Mob *spelltar,
 		int reflect_effectiveness,
 		bool use_resist_adjust,
@@ -5988,7 +5988,7 @@ bool Bot::SpellOnTarget(
 	return false;
 }
 
-bool Bot::IsImmuneToSpell(uint16 spell_id, Mob *caster) {
+bool Bot::IsImmuneToSpell(int32 spell_id, Mob *caster) {
 	bool Result = false;
 	if (!caster)
 		return false;
@@ -6018,7 +6018,7 @@ bool Bot::IsImmuneToSpell(uint16 spell_id, Mob *caster) {
 	return Result;
 }
 
-bool Bot::DetermineSpellTargets(uint16 spell_id, Mob *&spell_target, Mob *&ae_center, CastAction_type &CastAction, EQ::spells::CastingSlot slot) {
+bool Bot::DetermineSpellTargets(int32 spell_id, Mob *&spell_target, Mob *&ae_center, CastAction_type &CastAction, EQ::spells::CastingSlot slot) {
 	bool Result = false;
 	SpellTargetType target_type = spells[spell_id].target_type;
 
@@ -6034,7 +6034,7 @@ bool Bot::DetermineSpellTargets(uint16 spell_id, Mob *&spell_target, Mob *&ae_ce
 	return Result;
 }
 
-bool Bot::DoCastSpell(uint16 spell_id, uint16 target_id, EQ::spells::CastingSlot slot, int32 cast_time, int32 mana_cost, uint32* oSpellWillFinish, uint32 item_slot, uint32 aa_id) {
+bool Bot::DoCastSpell(int32 spell_id, uint16 target_id, EQ::spells::CastingSlot slot, int32 cast_time, int32 mana_cost, uint32* oSpellWillFinish, uint32 item_slot, uint32 aa_id) {
 	bool Result = false;
 	if (GetClass() == Class::Bard)
 		cast_time = 0;
@@ -6158,7 +6158,7 @@ void Bot::GenerateSpecialAttacks() {
 		SetSpecialAbility(SpecialAbility::TripleAttack, 1);
 }
 
-bool Bot::DoFinishedSpellSingleTarget(uint16 spell_id, Mob* spellTarget, EQ::spells::CastingSlot slot, bool& stopLogic) {
+bool Bot::DoFinishedSpellSingleTarget(int32 spell_id, Mob* spellTarget, EQ::spells::CastingSlot slot, bool& stopLogic) {
 
 	if (
 		spellTarget &&
@@ -6252,7 +6252,7 @@ bool Bot::DoFinishedSpellSingleTarget(uint16 spell_id, Mob* spellTarget, EQ::spe
 	return true;
 }
 
-bool Bot::DoFinishedSpellGroupTarget(uint16 spell_id, Mob* spellTarget, EQ::spells::CastingSlot slot, bool& stopLogic) {
+bool Bot::DoFinishedSpellGroupTarget(int32 spell_id, Mob* spellTarget, EQ::spells::CastingSlot slot, bool& stopLogic) {
 	bool isMainGroupMGB = false;
 
 	if (isMainGroupMGB && (GetClass() != Class::Bard)) {
@@ -7688,7 +7688,7 @@ void EntityList::ShowSpawnWindow(Client* client, int Distance, bool NamedOnly) {
 	return;
 }
 
-uint8 Bot::GetNumberNeedingHealedInGroup(Mob* tar, uint16 spell_type, uint16 spell_id, float range) {
+uint8 Bot::GetNumberNeedingHealedInGroup(Mob* tar, uint16 spell_type, int32 spell_id, float range) {
 	if (!TargetValidation(tar)) {
 		return 0;
 	}
@@ -7871,7 +7871,7 @@ bool Bot::GetNeedsHateRedux(Mob *tar) {
 	return false;
 }
 
-bool Bot::HasOrMayGetAggro(bool sit_aggro, uint32 spell_id) {
+bool Bot::HasOrMayGetAggro(bool sit_aggro, int32 spell_id) {
     if (
         !GetTarget() ||
         !GetTarget()->GetHateTop()
@@ -7950,7 +7950,7 @@ void Bot::RaidGroupSay(const char* msg, ...) {
 	}
 }
 
-bool Bot::UseDiscipline(uint32 spell_id, uint32 target) {
+bool Bot::UseDiscipline(int32 spell_id, uint32 target) {
 	if (!IsValidSpell(spell_id)) {
 		RaidGroupSay("Not a valid spell.");
 		return false;
@@ -8270,7 +8270,7 @@ void Bot::SetBotEnforceSpellSetting(bool enforceSpellSettings)
 	AI_AddBotSpells(GetBotSpellID());
 }
 
-bool Bot::AddBotSpellSetting(uint16 spell_id, BotSpellSetting* bs)
+bool Bot::AddBotSpellSetting(int32 spell_id, BotSpellSetting* bs)
 {
 	if (!IsValidSpell(spell_id) || !bs) {
 		return false;
@@ -8300,7 +8300,7 @@ bool Bot::AddBotSpellSetting(uint16 spell_id, BotSpellSetting* bs)
 	return true;
 }
 
-bool Bot::DeleteBotSpellSetting(uint16 spell_id)
+bool Bot::DeleteBotSpellSetting(int32 spell_id)
 {
 	if (!IsValidSpell(spell_id)) {
 		return false;
@@ -8323,7 +8323,7 @@ bool Bot::DeleteBotSpellSetting(uint16 spell_id)
 	return true;
 }
 
-BotSpellSetting* Bot::GetBotSpellSetting(uint16 spell_id)
+BotSpellSetting* Bot::GetBotSpellSetting(int32 spell_id)
 {
 	if (!IsValidSpell(spell_id) || !bot_spell_settings.count(spell_id)) {
 		return nullptr;
@@ -8464,7 +8464,7 @@ void Bot::LoadBotSpellSettings()
 	}
 }
 
-bool Bot::UpdateBotSpellSetting(uint16 spell_id, BotSpellSetting* bs)
+bool Bot::UpdateBotSpellSetting(int32 spell_id, BotSpellSetting* bs)
 {
 	if (!IsValidSpell(spell_id) || !bs) {
 		return false;
@@ -8669,7 +8669,7 @@ int32 Bot::GetRawItemAC()
 	return Total;
 }
 
-void Bot::SendSpellAnim(uint16 target_id, uint16 spell_id)
+void Bot::SendSpellAnim(uint16 target_id, int32 spell_id)
 {
 	if (!target_id || !IsValidSpell(spell_id)) {
 		return;
@@ -8866,7 +8866,7 @@ void Bot::AddBotStartingItems(uint16 race_id, uint8 class_id)
 	}
 }
 
-void Bot::SetSpellRecastTimer(uint16 spell_id, int32 recast_delay) {
+void Bot::SetSpellRecastTimer(int32 spell_id, int32 recast_delay) {
 	if (!IsValidSpell(spell_id)) {
 		OwnerMessage("Failed to set spell recast timer.");
 		return;
@@ -8911,7 +8911,7 @@ void Bot::SetSpellRecastTimer(uint16 spell_id, int32 recast_delay) {
 	}
 }
 
-uint32 Bot::GetSpellRecastTimer(uint16 spell_id)
+uint32 Bot::GetSpellRecastTimer(int32 spell_id)
 {
 	uint32 result = 0;
 
@@ -8944,7 +8944,7 @@ uint32 Bot::GetSpellRecastTimer(uint16 spell_id)
 	return result;
 }
 
-uint32 Bot::GetSpellRecastRemainingTime(uint16 spell_id)
+uint32 Bot::GetSpellRecastRemainingTime(int32 spell_id)
 {
 	uint32 result = 0;
 
@@ -8955,7 +8955,7 @@ uint32 Bot::GetSpellRecastRemainingTime(uint16 spell_id)
 	return result;
 }
 
-bool Bot::CheckSpellRecastTimer(uint16 spell_id)
+bool Bot::CheckSpellRecastTimer(int32 spell_id)
 {
 	ClearExpiredTimers();
 
@@ -8971,7 +8971,7 @@ bool Bot::CheckSpellRecastTimer(uint16 spell_id)
 	return false;
 }
 
-void Bot::SetDisciplineReuseTimer(uint16 spell_id, int32 reuse_timer)
+void Bot::SetDisciplineReuseTimer(int32 spell_id, int32 reuse_timer)
 {
 	if (!IsValidSpell(spell_id)) {
 		OwnerMessage("Failed to set discipline reuse timer.");
@@ -9017,7 +9017,7 @@ void Bot::SetDisciplineReuseTimer(uint16 spell_id, int32 reuse_timer)
 	}
 }
 
-uint32 Bot::GetDisciplineReuseTimer(uint16 spell_id)
+uint32 Bot::GetDisciplineReuseTimer(int32 spell_id)
 {
 	uint32 result = 0;
 
@@ -9045,7 +9045,7 @@ uint32 Bot::GetDisciplineReuseTimer(uint16 spell_id)
 	return result;
 }
 
-uint32 Bot::GetDisciplineReuseRemainingTime(uint16 spell_id) {
+uint32 Bot::GetDisciplineReuseRemainingTime(int32 spell_id) {
 	uint32 result = 0;
 
 	if (GetDisciplineReuseTimer(spell_id) > Timer::GetCurrentTime()) {
@@ -9055,7 +9055,7 @@ uint32 Bot::GetDisciplineReuseRemainingTime(uint16 spell_id) {
 	return result;
 }
 
-bool Bot::CheckDisciplineReuseTimer(uint16 spell_id)
+bool Bot::CheckDisciplineReuseTimer(int32 spell_id)
 {
 	ClearExpiredTimers();
 
@@ -9189,7 +9189,7 @@ uint32 Bot::GetItemReuseRemainingTime(uint32 item_id)
 	return result;
 }
 
-uint32 Bot::CalcSpellRecastTimer(uint16 spell_id)
+uint32 Bot::CalcSpellRecastTimer(int32 spell_id)
 {
 	uint32 result = 0;
 
@@ -9206,7 +9206,7 @@ uint32 Bot::CalcSpellRecastTimer(uint16 spell_id)
 	return result;
 }
 
-void Bot::ClearDisciplineReuseTimer(uint16 spell_id)
+void Bot::ClearDisciplineReuseTimer(int32 spell_id)
 {
 	if (spell_id && !IsValidSpell(spell_id)) {
 		OwnerMessage(
@@ -9283,7 +9283,7 @@ void Bot::ClearItemReuseTimer(uint32 item_id)
 	ClearExpiredTimers();
 }
 
-void Bot::ClearSpellRecastTimer(uint16 spell_id)
+void Bot::ClearSpellRecastTimer(int32 spell_id)
 {
 	if (spell_id && !IsValidSpell(spell_id)) {
 		OwnerMessage(
@@ -9546,7 +9546,7 @@ bool Bot::PrecastChecks(Mob* tar, uint16 spell_type) {
 	return true;
 }
 
-bool Bot::CastChecks(uint16 spell_id, Mob* tar, uint16 spell_type, bool prechecks, bool ae_check) {
+bool Bot::CastChecks(int32 spell_id, Mob* tar, uint16 spell_type, bool prechecks, bool ae_check) {
 	if (prechecks) {
 		if (!tar || tar->GetAppearance() == eaDead  || tar->GetHP() < 0) {
 			LogBotSpellChecksDetail("{} says, 'Cancelling cast due to CastChecks !tar.'", GetCleanName());
@@ -9866,7 +9866,7 @@ bool Bot::CastChecks(uint16 spell_id, Mob* tar, uint16 spell_type, bool precheck
 	return true;
 }
 
-bool Bot::CanCastSpellType(uint16 spell_type, uint16 spell_id, Mob* tar) {
+bool Bot::CanCastSpellType(uint16 spell_type, int32 spell_id, Mob* tar) {
 	if (!spell_id || !tar) {
 		LogBotSpellChecksDetail("{} says, 'Cancelling cast of {} on {} due to failsafe checks.'", GetCleanName(), (spell_id ? GetSpellName(spell_id) : (spell_type ? GetSpellTypeNameByID(spell_type) : "Unknown")), (tar ? tar->GetCleanName() : "Unknown"));
 		return false;
@@ -10081,7 +10081,7 @@ bool Bot::CanCastSpellType(uint16 spell_type, uint16 spell_id, Mob* tar) {
 	return true;
 }
 
-bool Bot::BotHasEnoughMana(uint16 spell_id) {
+bool Bot::BotHasEnoughMana(int32 spell_id) {
 	if (!IsValidSpell(spell_id)) {
 		return false;
 	}
@@ -10095,7 +10095,7 @@ bool Bot::BotHasEnoughMana(uint16 spell_id) {
 	return true;
 }
 
-bool Bot::IsTargetAlreadyReceivingSpell(Mob* tar, uint16 spell_id) {
+bool Bot::IsTargetAlreadyReceivingSpell(Mob* tar, int32 spell_id) {
 	if (!tar || !spell_id) {
 		return true;
 	}
@@ -10126,7 +10126,7 @@ bool Bot::IsTargetAlreadyReceivingSpell(Mob* tar, uint16 spell_id) {
 	return false;
 }
 
-bool Bot::DoResistCheck(Mob* tar, uint16 spell_id, int32 resist_limit) {
+bool Bot::DoResistCheck(Mob* tar, int32 spell_id, int32 resist_limit) {
 
 	if (!tar || spell_id == 0) {
 		return false;
@@ -10173,7 +10173,7 @@ bool Bot::DoResistCheck(Mob* tar, uint16 spell_id, int32 resist_limit) {
 	return true;
 }
 
-bool Bot::DoResistCheckBySpellType(Mob* tar, uint16 spell_id, uint16 spell_type) {
+bool Bot::DoResistCheckBySpellType(Mob* tar, int32 spell_id, uint16 spell_type) {
 	if (!tar || !IsValidSpell(spell_id)) {
 		return false;
 	}
@@ -10185,7 +10185,7 @@ bool Bot::DoResistCheckBySpellType(Mob* tar, uint16 spell_id, uint16 spell_type)
 	return DoResistCheck(tar, spell_id, GetSpellTypeResistLimit(spell_type));
 }
 
-bool Bot::IsValidTargetType(uint16 spell_id, int target_type, uint8 body_type) {
+bool Bot::IsValidTargetType(int32 spell_id, int target_type, uint8 body_type) {
 	if (!spell_id) {
 		return false;
 	}
@@ -10294,7 +10294,7 @@ bool Bot::IsMobEngagedByAnyone(Mob* tar) {
     return false;
 }
 
-bool Bot::IsValidMezTarget(Mob* owner, Mob* npc, uint16 spell_id) {
+bool Bot::IsValidMezTarget(Mob* owner, Mob* npc, int32 spell_id) {
 	if (npc->GetSpecialAbility(SpecialAbility::MesmerizeImmunity)) {
 		return false;
 	}
@@ -11326,7 +11326,7 @@ bool Bot::AttemptAICastSpell(uint16 spell_type, Mob* tar) {
 	return result;
 }
 
-bool Bot::AttemptAACastSpell(Mob* tar, uint16 spell_id, AA::Rank* rank) {
+bool Bot::AttemptAACastSpell(Mob* tar, int32 spell_id, AA::Rank* rank) {
 	if (!IsValidSpell(spell_id)) {
 		return false;
 	}
@@ -11431,7 +11431,7 @@ bool Bot::AttemptAACastSpell(Mob* tar, uint16 spell_id, AA::Rank* rank) {
 	return true;
 }
 
-bool Bot::AttemptForcedCastSpell(Mob* tar, uint16 spell_id, bool is_disc) {
+bool Bot::AttemptForcedCastSpell(Mob* tar, int32 spell_id, bool is_disc) {
 	if (!IsValidSpell(spell_id)) {
 		return false;
 	}
@@ -11636,7 +11636,7 @@ bool Bot::IsValidBotSpellType(uint16 spell_type) {
 	);
 }
 
-bool Bot::IsValidSpellTypeBySpellID(uint16 spell_type, uint16 spell_id) {
+bool Bot::IsValidSpellTypeBySpellID(uint16 spell_type, int32 spell_id) {
 	if (IsAEBotSpellType(spell_type) && !IsAnyAESpell(spell_id)) {
 		return false;
 	}
@@ -12249,7 +12249,7 @@ bool Bot::HasRequiredLoSForPositioning(Mob* tar) {
 	return true;
 }
 
-bool Bot::HasValidAETarget(Bot* caster, uint16 spell_id, uint16 spell_type, Mob* tar) {
+bool Bot::HasValidAETarget(Bot* caster, int32 spell_id, uint16 spell_type, Mob* tar) {
 	int spell_range = caster->GetActSpellRange(spell_id, spells[spell_id].range);
 	int spell_ae_range = caster->GetAOERange(spell_id);
 	int target_count = 0;
@@ -12474,7 +12474,7 @@ bool Bot::ValidStateCheck(Mob* other, bool same_raid_group) {
 	return true;
 }
 
-bool Bot::IsValidSpellTypeSubType(uint16 spell_type, uint16 sub_type, uint16 spell_id) {
+bool Bot::IsValidSpellTypeSubType(uint16 spell_type, uint16 sub_type, int32 spell_id) {
 	if (sub_type == UINT16_MAX) {
 		return true;
 	}
@@ -12586,8 +12586,8 @@ bool Bot::IsValidSpellTypeSubType(uint16 spell_type, uint16 sub_type, uint16 spe
 	return false;
 }
 
-uint16 Bot::GetSpellByAA(int id, AA::Rank*& rank) {
-	uint16 spell_id = 0;
+int32 Bot::GetSpellByAA(int id, AA::Rank*& rank) {
+	int32 spell_id = 0;
 	std::pair<AA::Ability*, AA::Rank*> aa_ability = std::make_pair(nullptr, nullptr);
 	AA::Ability* ability = zone->GetAlternateAdvancementAbility(id);
 
@@ -12612,7 +12612,7 @@ uint16 Bot::GetSpellByAA(int id, AA::Rank*& rank) {
 	return spell_id;
 }
 
-void Bot::SetBotBlockedBuff(uint16 spell_id, bool block)
+void Bot::SetBotBlockedBuff(int32 spell_id, bool block)
 {
 	if (!IsValidSpell(spell_id)) {
 		OwnerMessage("Failed to set blocked buff.");
@@ -12659,7 +12659,7 @@ bool Bot::IsBlockedBuff(int32 spell_id)
 	return result;
 }
 
-void Bot::SetBotBlockedPetBuff(uint16 spell_id, bool block)
+void Bot::SetBotBlockedPetBuff(int32 spell_id, bool block)
 {
 	if (!IsValidSpell(spell_id)) {
 		OwnerMessage("Failed to set blocked pet buff.");
@@ -13319,7 +13319,7 @@ uint8 Bot::GetUltimateSpellTypeMaxThreshold(uint16 spell_type, Mob* tar) {
 	return GetSpellTypeMaxThreshold(spell_type);
 }
 
-bool Bot::IsImmuneToBotSpell(uint16 spell_id, Mob* caster) {
+bool Bot::IsImmuneToBotSpell(int32 spell_id, Mob* caster) {
 	int effect_index;
 
 	if (!caster) {
